+++ src/accesodatos/RegistroParqueoDAO.java
@@ -10,9 +10,15 @@ public interface RegistroParqueoDAO {
 
     List<RegistroParqueo> obtenerRegistrosActivos() throws IOException;
 
+    List<RegistroParqueo> obtenerRegistrosHistorial() throws IOException;
+
     List<RegistroParqueo> obtenerTodosLosRegistros() throws IOException;
 
     boolean existePlacaActiva(String placa) throws IOException;
+
+    RegistroParqueo buscarRegistroActivoPorId(int idRegistro) throws IOException;
+
+    void actualizarRegistro(RegistroParqueo registroActualizado) throws IOException;
 
     int obtenerSiguienteId() throws IOException;
 }